--- a/discovery-provider/src/queries/search_queries.py
+++ b/discovery-provider/src/queries/search_queries.py
@@ -1,4 +1,5 @@
 from enum import Enum
+import concurrent.futures
 import logging  # pylint: disable=C0302
 from functools import cmp_to_key
 from flask import Blueprint, request
@@ -218,6 +219,64 @@ def add_users(session, results):
             result["user"] = user
     return results
 
+def perform_search_query(session, search_type, args):
+    search_str = args.get('search_str')
+    limit = args.get('limit')
+    offset = args.get('offset')
+    is_auto_complete = args.get('is_auto_complete')
+    current_user_id = args.get('current_user_id')
+    only_downloadable = args.get('only_downloadable')
+
+    results = []
+    if search_type == 'tracks':
+        results = track_search_query(session, search_str, limit, offset, False, is_auto_complete, current_user_id, only_downloadable)
+    elif search_type == 'saved_tracks':
+        results = track_search_query(
+            session, search_str, limit, offset, True, is_auto_complete, current_user_id, only_downloadable)
+    elif search_type == 'users':
+        results = user_search_query(
+            session, search_str, limit, offset, False, is_auto_complete, current_user_id)
+    elif search_type == 'followed_users':
+        results = user_search_query(
+            session, search_str, limit, offset, True, is_auto_complete, current_user_id)
+    elif search_type == 'playlists':
+        results = playlist_search_query(
+            session,
+            search_str,
+            limit,
+            offset,
+            False,
+            False,
+            is_auto_complete,
+            current_user_id
+        )
+    elif search_type == 'saved_playlists':
+        results = playlist_search_query(
+            session,
+            search_str,
+            limit,
+            offset,
+            False,
+            True,
+            is_auto_complete,
+            current_user_id
+        )
+    elif search_type == 'albums':
+        results = playlist_search_query(
+            session, search_str, limit, offset, True, False, is_auto_complete, current_user_id)
+    elif search_type == 'saved_albums':
+        results = playlist_search_query(
+            session,
+            search_str,
+            limit,
+            offset,
+            True,
+            True,
+            is_auto_complete,
+            current_user_id
+        )
+    return results
+
 # SEARCH QUERIES
 # We chose to use the raw SQL instead of SQLAlchemy because we're pushing SQLAlchemy to it's
 # limit to do this query by creating new wrappers for pg functions that do not exist like
@@ -258,88 +317,80 @@ def search(args):
     searchKind = SearchKind[kind]
 
     results = {}
+
+    # Accumulate user_ids for later
+    user_ids = set()
+
+    # Create args for perform_search_query
+    search_args = {
+        "search_str": searchStr,
+        "limit": limit,
+        "offset": offset,
+        "is_auto_complete": is_auto_complete,
+        "current_user_id": current_user_id,
+        "only_downloadable": only_downloadable
+    }
+
     if searchStr:
         db = get_db_read_replica()
         with db.scoped_session() as session:
-            def with_users_added(results):
-                if not with_users:
-                    return results
-                return add_users(session, results)
+            with concurrent.futures.ThreadPoolExecutor(max_workers=5) as executor:
+                # Keep a mapping of future -> search_type
+                futures_map = {}
+                futures = []
 
-            # Accumulate user_ids for later
-            user_ids = set()
+                # Helper fn to submit a future and add it to bookkeeping data structures
+                def submit_and_add(search_type):
+                    future = executor.submit(perform_search_query, session, search_type, args)
+                    futures.append(future)
+                    futures_map[future] = search_type
 
-            if (searchKind in [SearchKind.all, SearchKind.tracks]):
-                results['tracks'] = track_search_query(
-                    session, searchStr, limit, offset, False, is_auto_complete, current_user_id, only_downloadable)
-                user_ids.update(get_users_ids(results['tracks']))
-                if current_user_id:
-                    results['saved_tracks'] = track_search_query(
-                        session, searchStr, limit, offset, True, is_auto_complete, current_user_id, only_downloadable)
-                    user_ids.update(get_users_ids(results['saved_tracks']))
-            if (searchKind in [SearchKind.all, SearchKind.users]):
-                results['users'] = user_search_query(
-                    session, searchStr, limit, offset, False, is_auto_complete, current_user_id)
-                if current_user_id:
-                    results['followed_users'] = user_search_query(
-                        session, searchStr, limit, offset, True, is_auto_complete, current_user_id)
-            if (searchKind in [SearchKind.all, SearchKind.playlists]):
-                results['playlists'] = playlist_search_query(
-                    session,
-                    searchStr,
-                    limit,
-                    offset,
-                    False,
-                    False,
-                    is_auto_complete,
-                    current_user_id
-                )
-                user_ids.update(get_users_ids(results['playlists']))
-                if current_user_id:
-                    results['saved_playlists'] = playlist_search_query(
-                        session,
-                        searchStr,
-                        limit,
-                        offset,
-                        False,
-                        True,
-                        is_auto_complete,
-                        current_user_id
-                    )
-                    user_ids.update(get_users_ids(results['saved_playlists']))
-            if (searchKind in [SearchKind.all, SearchKind.albums]):
-                results['albums'] = playlist_search_query(
-                    session, searchStr, limit, offset, True, False, is_auto_complete, current_user_id)
-                user_ids.update(get_users_ids(results['albums']))
-                if current_user_id:
-                    results['saved_albums'] = playlist_search_query(
-                        session,
-                        searchStr,
-                        limit,
-                        offset,
-                        True,
-                        True,
-                        is_auto_complete,
-                        current_user_id
-                    )
-                user_ids.update(get_users_ids(results['albums']))
+                if (searchKind in [SearchKind.all, SearchKind.tracks]):
+                    submit_and_add("tracks")
+                    if current_user_id:
+                        submit_and_add("saved_tracks")
 
-            # Add users back
-            users = get_users_by_id(session, list(user_ids), current_user_id)
-            logger.warning("USERS MAP!")
+                if (searchKind in [SearchKind.all, SearchKind.users]):
+                    submit_and_add("users")
+                    if current_user_id:
+                        submit_and_add("followed_users")
 
-            for (_, result_list) in results.items():
-                for result in result_list:
-                    user_id = None
-                    if 'playlist_owner_id' in result:
-                        user_id = result['playlist_owner_id']
-                    elif 'owner_id' in result:
-                        user_id = result['owner_id']
+                if (searchKind in [SearchKind.all, SearchKind.playlists]):
+                    submit_and_add("playlists")
+                    if current_user_id:
+                        submit_and_add("saved_playlists")
 
-                    if user_id is not None:
-                        user = users[user_id]
-                        result["user"] = user
+                if (searchKind in [SearchKind.all, SearchKind.albums]):
+                    submit_and_add("albums")
+                    if current_user_id:
+                        submit_and_add("saved_albums")
 
+                # TODO: add a timeout here
+                for future in concurrent.futures.as_completed(futures):
+                    search_result = future.result()
+                    future_type = futures_map[future]
+
+                    # Add to the final results
+                    results[future_type] = search_result
+
+                    # Add to user_ids
+                    user_ids.update(get_users_ids(search_result))
+
+                # Add users back
+                users = get_users_by_id(session, list(user_ids), current_user_id)
+
+                for (_, result_list) in results.items():
+                    for result in result_list:
+                        user_id = None
+                        if 'playlist_owner_id' in result:
+                            user_id = result['playlist_owner_id']
+                        elif 'owner_id' in result:
+                            user_id = result['owner_id']
+
+                        if user_id is not None:
+                            user = users[user_id]
+                            result["user"] = user
+    logger.warning(f"{results}")
     return results
 
 def track_search_query(