### Description
Improves perf by:
- Moving requesting users to the end of the computation, essentially halving the # of db round trips
- Parallelizes different search type DB lookups as much as possible. 

Results:
- Regular search is > 2x faster (~5.5s -> 2.75s)
- Autocomplete search is ~30% faster (1.5s -> 1s)

And this will get even better after the populate fixes are in.

There's more we can do to improve search, but let's get some of the low hanging fruit in before spending time figuring out the next line of attack.

### Tests

Tested on my remote prod machine.