@piazzatron 2021-04-14
> Change looks good for futures, etc. Curious to see how this does play out in the live prod db. I did notice that we are always including users with this change, which maybe is adding extra payload to the api search queries. Maybe I'm missing something though?

Greeaaat call. Gonna to make sure we're not including that for regular API.